Create README.md
Hey there! I'm Raha, and I'm thrilled to be diving into the world of data analysis on github! With a passion for economics and a love for unraveling the stories hidden within data, I'm here to bring a fresh perspective to your projects. My journey through academia and research has equipped me with a solid foundation, but it's my hunger for discovery and my knack for turning complex datasets into actionable insights that truly sets me apart.